backend/requirements.txt
# --- Web API ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# --- Env / config ---
python-dotenv>=1.0.1

# --- Validation / email ---
pydantic>=2.7.0
email-validator>=2.2.0

# --- MongoDB ---
pymongo>=4.7.0

# --- Auth / JWT / Password hashing ---
PyJWT>=2.9.0
pwdlib[argon2]>=0.2.0

# --- File uploads (multipart/form-data) ---
python-multipart>=0.0.9

# --- Background tasks ---
celery>=5.4.0
redis>=5.0.0

# --- Scraping ---
requests>=2.32.0
beautifulsoup4>=4.12.3
lxml>=5.2.0

# --- Google Gmail API / OAuth ---
google-api-python-client>=2.140.0
google-auth>=2.32.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1

